Create new color instead of composite.  Switch to value from float (0-255 instead of 0-1).

+++ src/org/powerbot/script/xenon/wrappers/Drawable.java
@@ -5,5 +5,5 @@ import java.awt.Graphics;
 public interface Drawable {
 	public void draw(Graphics render);
 
-	public void draw(Graphics render, float alpha);
+	public void draw(Graphics render, int alpha);
 }
